+++ packages/core/menu/src/types.ts
@@ -9,19 +9,22 @@ export type SectionProps = {
   children: React.ReactNode;
 };
 
-export interface ButtonItemProps {
+export interface BaseItemProps {
   elemBefore?: React.ReactNode;
   elemAfter?: React.ReactNode;
   onClick?: (event: React.MouseEvent | React.KeyboardEvent) => void;
   title?: string;
   description?: string;
   isDisabled?: boolean;
-  children: React.ReactNode;
+  isSelected?: boolean;
+  children?: React.ReactNode;
+  href?: string;
+  component?: React.ComponentType<{ wrapperClass: string }>;
 }
 
-export interface LinkItemProps extends ButtonItemProps {
-  href: string;
-}
+export type ButtonItemProps = Omit<BaseItemProps, 'component' | 'href'>;
+export type LinkItemProps = Omit<BaseItemProps, 'component'>;
+export type CustomItemProps = Omit<BaseItemProps, 'href'>;
 
 export type SkeletonItemProps = {
   hasAvatar?: boolean;