+++ packages/editor/editor-wikimarkup-transformer/src/parser/tokenize/links/issue-link.ts
@@ -9,6 +9,16 @@ export function issueLinkResolver(
   context: Context,
 ): PMNode[] | undefined {
   const { originalLinkText, linkTitle, notLinkBody } = link;
+  if (
+    linkTitle === 'smart-card' ||
+    linkTitle === 'block-link' // TODO: Depricated should be removed in the next major release
+  ) {
+    return [
+      schema.nodes.blockCard.createChecked({
+        url: notLinkBody,
+      }),
+    ];
+  }
   if (linkTitle === 'smart-link') {
     return [
       schema.nodes.inlineCard.createChecked({