+++ packages/monorepo-tooling/branch-deploy-product-integrator/src/__tests__/unit/cli.ts
@@ -1,0 +1,15 @@
+describe('CLI', () => {
+  it.todo('<push> should call push fn with correct arguments');
+
+  it.todo(
+    '<report-status> should call report status fn with correct arguments',
+  );
+
+  describe('Help', () => {
+    it.todo('<help> should show help information');
+
+    it.todo('Should show push help information');
+
+    it.todo('Should show report-status help information');
+  });
+});