+++ packages/help/help/CHANGELOG.md
@@ -1,5 +1,37 @@
 ## 0.1.0
 
+## 0.21.4
+
+### Patch Changes
+
+- [patch][86198c3e5d](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/86198c3e5d):
+
+  Bugfix - Fix voice-over accessibility issues
+
+## 0.21.3
+
+### Patch Changes
+
+- [patch][da00b546cb](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/da00b546cb):
+
+  Bugfix - Hide DefaultContent if an article is fully visible
+
+## 0.21.2
+
+### Patch Changes
+
+- [patch][bb95b5ff7e](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/bb95b5ff7e):
+
+  Updated dependency help-article
+
+## 0.21.1
+
+### Patch Changes
+
+- [patch][bd17edc20c](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/bd17edc20c):
+
+  Bugfix - hide back button if therere is only one article in the history and there isn't default content
+
 ## 0.21.0
 
 ### Minor Changes