+++ packages/navigation/atlassian-navigation/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @atlaskit/atlassian-navigation
 
+## 0.6.5
+
+### Patch Changes
+
+- [patch][e419c3c01f](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/e419c3c01f):
+
+  Update color of skeleton in atlassian-navigation
+
 ## 0.6.4
 
 ### Patch Changes